requirements.txt
telethon>=1.35.0
python-telegram-bot>=20.0
nest_asyncio
python-dotenv==1.0.0